+++ docs/changes.rst
@@ -4,6 +4,12 @@ Wand Changelog
 0.4 series
 ~~~~~~~~~~
 
+Version 0.4.4
+~~~~~~~~~~~~~
+
+To be released.
+
+
 Version 0.4.3
 -------------